+++ lib/core/plugins.py
