--- plugins/CVE-2021-3156.py
@@ -1,25 +1,0 @@
-# Commands to be executed on the target
-COMMANDS = [
-    'sudoedit -s "1234567890123456789012\\" 2>&1',
-    "sudo --version 2>&1"
-]
-# Matcher string for distribution/*nix
-# https://docs.python.org/3.4/library/re.html
-MATCHER = ".*"
-
-def get_commands() -> list:
-    return COMMANDS
-
-
-def get_matcher() -> str:
-    return MATCHER
-
-
-def parse(output: list) -> str:
-    if "memory" in output[0] or "Error" in output[0] or "Backtrace" in output[0]:
-        retVal = "sudo -V\n"
-        retVal += f"{output[1]}\n"
-        retVal += "Is vulnerable to CVE-2021-3156"
-        return retVal
-    
-    return "Not vulnerable"
