+++ plugins/basic.yaml
