+++ plugins/python2.yaml
